+++ Cargo.toml
@@ -31,7 +31,7 @@ rtcp = "0.7.1"
 cpal = "0.15.2"
 iced = { version = "0.9.0", features = ["tokio", "image", "svg"] }
 iced_aw = { version = "0.5.2", features = ["tabs"] }
-apple-sys = { version = "0.2.0", features = ["ScreenCaptureKit"] }
+apple-sys = { version = "0.2.0", features = ["CoreMedia", "ScreenCaptureKit"] }
 objc = "0.2.3"
 
 [target.'cfg(target_os = "macos")'.dependencies]
