+++ src/capture/macos/mod.rs
@@ -1,10 +1,7 @@
 pub use macos_capture::MacOSScreenCapture;
 
-mod config;
+mod capture_engine;
 pub mod display;
 mod ffi;
 pub mod macos_capture;
-pub mod macos_sc_capture;
-mod capture_engine;
-mod ffi_sc;
-
+pub mod screen_recorder;
